Refactor web example
