+++ example/common/helpers/index.js
