+++ example/native/android/app/src/main/java/com/example/MainApplication.java
