+++ example/native/android/build.gradle
