+++ example/native/android/gradle.properties
