+++ example/native/android/settings.gradle
