+++ example/native/app.json
