+++ example/native/ios/example-tvOSTests/Info.plist
