+++ example/native/ios/example.xcodeproj/xcshareddata/xcschemes/example.xcscheme
