+++ example/native/ios/example/Images.xcassets/AppIcon.appiconset/Contents.json
