+++ example/native/ios/exampleTests/Info.plist
