+++ example/native/package.json
@@ -1,0 +1,16 @@
+{
+	"name": "example",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"start": "node node_modules/react-native/local-cli/cli.js start",
+		"test": "jest"
+	},
+	"dependencies": {
+		"react-native": "0.43.2",
+		"react-native-payments": "file:..",
+	},
+	"jest": {
+		"preset": "react-native"
+	}
+}
