--- example/package.json
@@ -1,34 +1,0 @@
-{
-	"name": "example",
-	"version": "0.0.1",
-	"private": true,
-	"scripts": {
-		"start": "node node_modules/react-native/local-cli/cli.js start",
-		"test": "jest",
-		"web:start": "webpack-dev-server --open --config=web/webpack.development.config.js",
-		"web:build": "webpack --config=web/webpack.production.config.js --color --progress",
-		"web:deploy": "now ./web/public/ --name='rnp-web' --public",
-		"deploy": "yarn run web:build && yarn run web:deploy"
-	},
-	"dependencies": {
-		"react": "^15.6.1",
-		"react-dom": "^15.6.1",
-		"react-native": "0.43.2",
-		"react-native-payments": "file:..",
-		"react-primitives": "^0.4.2"
-	},
-	"devDependencies": {
-		"babel-core": "^6.25.0",
-		"babel-jest": "19.0.0",
-		"babel-loader": "^7.0.0",
-		"babel-preset-env": "^1.5.2",
-		"babel-preset-react-native": "1.9.1",
-		"jest": "19.0.2",
-		"react-test-renderer": "16.0.0-alpha.6",
-		"webpack": "^2.6.1",
-		"webpack-dev-server": "^2.4.5"
-	},
-	"jest": {
-		"preset": "react-native"
-	}
-}
