+++ lib/js/index.js
@@ -1,3 +1,3 @@
 import PaymentRequest from './PaymentRequest';
 
-export default { PaymentRequest }
+module.exports = { PaymentRequest };
